incr-12

+++ internal/store/database.go
@@ -10,7 +10,7 @@ import (
 var schema = `CREATE TABLE metrics (
 				  id VARCHAR(256) CONSTRAINT metric_id_pk PRIMARY KEY NOT NULL,
 				  metric_type VARCHAR(256) NOT NULL,
-				  counter_value INTEGER DEFAULT NULL,
+				  counter_value NUMERIC DEFAULT NULL,
 				  gauge_value DOUBLE PRECISION DEFAULT NULL,
 				  hash VARCHAR(256) DEFAULT NULL);`
